Add files via upload
As an adaption of my Pyra command line audio/video editor, I wanted a version that makes it easier to use. The main point of the program is to bring yt-dlp with a graphical user interface, and allow the user to download a video on the fly and user common editing tools on the video in one go.